pyproject.toml
[project]
name = "ace"
version = "0.1.0"
description = "Adaptive Chat Extractor - Une solution universelle pour l'extraction de conversations HTML."
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4",
    "PyYAML",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
ace = "ace.cli:main"

[tool.setuptools]
packages = ["ace"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
